resolve issue with spring() and push() / pop()
resolves #22

After investigating, it seems that the original way I was handling the
spring function being updated one the values[] array change was not
accounting for new values in the array. so the spring function would
only apply to the original values, and no new ones.

Resolved this by explicitly re-setting the spring function values after
detecting a change in the length of the array, but only applying a
.set() function if the length has not changed.