change the way darkmode works
i realised that setting the darkmode with automatic @media would cause a
lot of backwards incompatibility. So I have made darkmode an explicit
option with the default value of false. It can be opt'd in with 'force'
or 'auto' to either always be dark, or use the system preference
respectively.